+++ flusher.go
@@ -357,8 +357,13 @@ func (s *Server) flushForward(ctx context.Context, wms []WorkerMetrics) {
 
 	// the error has already been logged (if there was one), so we only care
 	// about the success case
-	if vhttp.PostHelper(span.Attach(ctx), s.HTTPClient, s.TraceClient, http.MethodPost, fmt.Sprintf("%s/import", s.ForwardAddr), jsonMetrics, "forward", true, log) == nil {
-		log.WithField("metrics", len(jsonMetrics)).Info("Completed forward to upstream Veneur")
+	endpoint := fmt.Sprintf("%s/import", s.ForwardAddr)
+	if vhttp.PostHelper(span.Attach(ctx), s.HTTPClient, s.TraceClient, http.MethodPost, endpoint, jsonMetrics, "forward", true, log) == nil {
+		log.WithFields(logrus.Fields{
+			"metrics":     len(jsonMetrics),
+			"endpoint":    endpoint,
+			"forwardAddr": s.ForwardAddr,
+		}).Info("Completed forward to upstream Veneur")
 	}
 }