+++ handlers_global.go
@@ -11,6 +11,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/sirupsen/logrus"
 	"github.com/stripe/veneur/samplers"
 	"github.com/stripe/veneur/ssf"
 	"github.com/stripe/veneur/trace"
@@ -26,8 +27,13 @@ func (ch contextHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 func handleProxy(p *Proxy) http.Handler {
 	return contextHandler(func(ctx context.Context, w http.ResponseWriter, r *http.Request) {
+		log.WithFields(logrus.Fields{
+			"path": r.URL.Path,
+			"host": r.URL.Host,
+		}).Debug("Importing metrics on proxy")
 		span, jsonMetrics, err := unmarshalMetricsFromHTTP(ctx, p.TraceClient, w, r)
 		if err != nil {
+			log.WithError(err).Error("Error unmarshalling metrics in proxy import")
 			return
 		}
 		// the server usually waits for this to return before finalizing the
@@ -55,6 +61,8 @@ func handleImport(s *Server) http.Handler {
 	return contextHandler(func(ctx context.Context, w http.ResponseWriter, r *http.Request) {
 		span, jsonMetrics, err := unmarshalMetricsFromHTTP(ctx, s.TraceClient, w, r)
 		if err != nil {
+			log.WithError(err).Error("Error unmarshalling metrics in global import")
+			span.Add(ssf.Count("import.unmarshal.errors_total", 1, nil))
 			return
 		}
 		// the server usually waits for this to return before finalizing the
@@ -136,7 +144,7 @@ func unmarshalMetricsFromHTTP(ctx context.Context, client *trace.Client, w http.
 			span.Error(err)
 			encLogger.WithError(err).Error("Could not read compressed request body")
 			span.Add(ssf.Count("import.request_error_total", 1, map[string]string{"cause": "deflate"}))
-			return nil, nil, err
+			return span, nil, err
 		}
 		defer body.Close()
 	default:
@@ -144,7 +152,7 @@ func unmarshalMetricsFromHTTP(ctx context.Context, client *trace.Client, w http.
 		span.Error(errors.New("Could not determine content-encoding of request"))
 		encLogger.Error("Could not determine content-encoding of request")
 		span.Add(ssf.Count("import.request_error_total", 1, map[string]string{"cause": "unknown_content_encoding"}))
-		return nil, nil, err
+		return span, nil, err
 	}
 
 	if err = json.NewDecoder(body).Decode(&jsonMetrics); err != nil {
@@ -152,7 +160,7 @@ func unmarshalMetricsFromHTTP(ctx context.Context, client *trace.Client, w http.
 		span.Error(err)
 		innerLogger.WithError(err).Error("Could not decode /import request")
 		span.Add(ssf.Count("import.request_error_total", 1, map[string]string{"cause": "json"}))
-		return nil, nil, err
+		return span, nil, err
 	}
 
 	if len(jsonMetrics) == 0 {
@@ -160,7 +168,7 @@ func unmarshalMetricsFromHTTP(ctx context.Context, client *trace.Client, w http.
 		http.Error(w, msg, http.StatusBadRequest)
 		span.Error(errors.New(msg))
 		innerLogger.WithError(err).Error(msg)
-		return nil, nil, err
+		return span, nil, err
 	}
 
 	// We want to make sure we have at least one entry