+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ** New type `ssf.Samples` holding a batch of samples which can be submitted conveniently through `trace/metrics`.
 ** Method `trace.(*Trace).Add`, which allows adding metrics to a trace span.
 * `veneur-proxy` has a new configuration option `forward_timeout` which allows specifying how long forwarding a batch to global veneur servers may take in total. Thanks, [antifuchs](https://github.com/antifuchs)!
+* Add native support for running Veneur within Kubernetes. Thanks, [aditya](https://github.com/chimeracoder)!
 
 ## Improvements
 * Updated Datadog span sink to latest version in Datadog tracing agent. Thanks, [gphat](https://github.com/gphat)!